+++ flagfit-lint/src/main/java/tv/abema/flagfit/AnnotationPackagePath.kt
@@ -1,0 +1,10 @@
+package tv.abema.flagfit
+
+object AnnotationPackagePath {
+  const val PACKAGE_PATH_WIP = "tv.abema.flagfit.FlagType.WorkInProgress"
+  const val PACKAGE_PATH_EXPERIMENT = "tv.abema.flagfit.FlagType.Experiment"
+  const val PACKAGE_PATH_OPS = "tv.abema.flagfit.FlagType.Ops"
+  const val PACKAGE_PATH_PERMISSION = "tv.abema.flagfit.FlagType.Permission"
+  const val PACKAGE_PATH_BOOLEAN_FLAG = "tv.abema.flagfit.annotation.BooleanFlag"
+  const val PACKAGE_PATH_VARIATION_FLAG = "tv.abema.flagfit.annotation.VariationFlag"
+}
