+++ flagfit-lint/src/main/java/tv/abema/flagfit/IssueRegistry.kt
@@ -22,6 +22,7 @@ internal class IssueRegistry : IssueRegistry() {
   override val issues: List<Issue>
     get() = listOf(
       DeadlineExpiredDetector.ISSUE_DEADLINE_EXPIRED,
-      DeadlineExpiredDetector.ISSUE_DEADLINE_SOON
+      DeadlineExpiredDetector.ISSUE_DEADLINE_SOON,
+      IllegalParamDetector.ISSUE_ILLEGAL_NO_EXPIRE_PARAM
     )
 }
